+++ src/message.ts
@@ -1,9 +1,6 @@
-export const LEGACY_VSCODE_LINTER_CONFIG_MESSAGE =
-  "You have legacy linter options in your VS Code settings. They are no longer being used. See [documentation for migration information](https://github.com/prettier/prettier-vscode#linter-integration). Would you like to remove them or view the logs?";
 export const OUTDATED_PRETTIER_VERSION_MESSAGE =
   "Your project is configured to use an outdated version of prettier that cannot be used by this extension. Upgrade to the latest version of prettier.";
 export const VIEW_LOGS_ACTION_TEXT = "Show Log";
-export const REMOVE_LEGACY_OPTIONS_ACTION_TEXT = "Remove Settings";
 export const FAILED_TO_LOAD_MODULE_MESSAGE =
   "Failed to load module. If you have prettier or plugins referenced in package.json, ensure you have run `npm install`";
 export const INVALID_PRETTIER_CONFIG =