+++ src/test/suite/notifications.test.ts
@@ -3,10 +3,7 @@ import * as assert from "assert";
 import * as sinon from "sinon";
 // tslint:disable-next-line: no-implicit-dependencies
 import { MessageItem, MessageOptions, window } from "vscode";
-import {
-  LEGACY_VSCODE_LINTER_CONFIG_MESSAGE,
-  OUTDATED_PRETTIER_VERSION_MESSAGE,
-} from "../../message";
+import { OUTDATED_PRETTIER_VERSION_MESSAGE } from "../../message";
 import { format } from "./format.test";
 
 suite("Test notifications", function () {
@@ -30,10 +27,6 @@ suite("Test notifications", function () {
   test("shows error for outdated prettier instance", async () => {
     await format("outdated", "ugly.js");
     assert(showErrorMessage.calledWith(OUTDATED_PRETTIER_VERSION_MESSAGE));
-  });
-  test("shows error for legacy vscode config", async () => {
-    await format("outdated", "ugly.js");
-    assert(showWarningMessage.calledWith(LEGACY_VSCODE_LINTER_CONFIG_MESSAGE));
   });
   test("does not show error with valid project", async () => {
     await format("plugins", "index.php");